+++ lib/src/board_controller.dart
@@ -35,7 +35,7 @@ class BoardController extends StatefulWidget {
   /// A vector to offset the position of dragged pieces by, relative to the size of the piece.
   /// No offset is recommended for web, and Offset(0,-1) for mobile, in which case
   /// the bottom of the piece will be anchored to the finger.
-  final Offset dragOffset;
+  final Offset dragFeedbackOffset;
 
   late final Map<int, List<Move>> moveMap;
   late final List<Move> drops;
@@ -63,7 +63,7 @@ class BoardController extends StatefulWidget {
     required this.canMove,
     this.draggable = true,
     this.dragFeedbackSize = 2.0,
-    this.dragOffset = const Offset(0.0, -1.0),
+    this.dragFeedbackOffset = const Offset(0.0, -1.0),
     this.allowAnimation = true,
     this.animationDuration,
     this.animationCurve,
@@ -316,7 +316,7 @@ class _BoardControllerState extends State<BoardController> {
           canMove: widget.canMove,
           draggable: widget.draggable,
           dragFeedbackSize: widget.dragFeedbackSize,
-          dragOffset: widget.dragOffset,
+          dragFeedbackOffset: widget.dragFeedbackOffset,
           onTap: onTap,
           onDragCancel: onDragCancel,
           validateDrag: validateDrag,